requirements.txt
asgiref==3.4.1
Django==3.2.5
django-admin-interface==0.16.3
django-colorfield==0.4.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==8.3.1
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
